backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
sentence-transformers==3.3.1
numpy==2.1.3
yfinance==0.2.51
reportlab==4.2.5
# Use CPU wheels on Windows unless you install CUDA PyTorch with enough disk space:
#   pip install torch --index-url https://download.pytorch.org/whl/cu124
torch>=2.2.0
transformers>=4.40.0
peft>=0.11.0
accelerate>=0.26.0
safetensors>=0.4.0
